requirements.txt
attrs==22.1.0
iniconfig==1.1.1
llvmlite==0.39.1
numpy==1.23.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pysha3==1.0.2
pytest==7.1.3
tomli==2.0.1
